tests/wpt/web-platform-tests/service-workers/service-worker/update-top-level.https.html

<!DOCTYPE html>
<title>Service Worker: Registration update()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';

function wait_for_message() {
  return new Promise(resolve => {
    navigator.serviceWorker.addEventListener("message",
      e => {
        resolve(e.data);
      }, { once: true });
  });
}

promise_test(async t => {
  const script = './resources/update-top-level-worker.py';
  const scope = './resources/empty.html?update-result';

  let reg = await navigator.serviceWorker.register(script, { scope });
  t.add_cleanup(async _ => await reg.unregister());
  await wait_for_state(t, reg.installing, 'activated');

  reg.addEventListener("updatefound",
    () => assert_unreached("shouldn't find an update"));

  reg.active.postMessage("ping");
  assert_equals(await wait_for_message(), 'PASS', 'did not hang');
}, 'A serviceworker with a top-level update should not hang');
</script>
